Rename components/SideNavigation.vue to components/_shared/SideNavigation.vue
